Bug 1051191 - Use source-built updater binary for FOTA packages r=gsvelto
We were using a prebuilt updater-binary, dating from old ICS days. This
is not optimal: we have no way to customize the FOTA process, newer SDK
version may not be compatible, and it makes us dependant on a piece of
blob. We solve this by adding a build dependency against the 'updater'
package, so that the updater binary is built. We then use it from
system/bin/updater to include it as the update-binary that is being used
during FOTA.