Crop the boot animations to speed up boot
We're wasting a crap ton of time decoding black, when we should be
taking advantage of the fact that the bootanimation service will
fill the background with black.

This also prevents boot animation jank because the initial loading
of each frame is now less than the 1/30s for each frame.

Signed-off-by: Adam Farden <adam@farden.cz>