Bug 1240352 - Allow blobfree shell scripts to use proprietary-files.txt r=gerard-mjax
The blobfree scripts assume that all blobs will be copied from a
blobs.mk but actually some are copied 'properly' in Android.mk.
This results in some blobs not used by the blobfree process.

Actually there are several methods a blob can be copied to $OUT
so we should use the proprietary-files.txt provided by CM repos
which is a list of ALL blobs extracted from a device. This is
used by extract-files.sh which pulls the blobs and creates the
various mk files in /vendor.

Instead of reading and processing the mk files directly we create
an intermediate file in $OUT that ensures the list is in the same
format no matter which type of blob list is used.

Several devices use the existing method so we default to this, and
use this new method with the TARGET_DEVICE_BLOBS_LIST flag.